pyproject.toml
[project]
name = "ami-browser"
version = "0.1.0"
description = "Browser module for AMI orchestrator"
readme = "README.md"
requires-python = ">=3.12"
authors = [ { name = "AMI Team" } ]

dependencies = [
  "pydantic==2.11.7",
  "loguru==0.7.3",
  "psutil==7.0.0",
  "opencv-python==4.12.0.88",
  "PyYAML==6.0.2",
  "Pillow==11.3.0",
  "beautifulsoup4==4.13.4",
  "aiofiles==24.1.0",
  "aiohttp==3.12.15",
  "webdriver-manager==4.0.2",
  "chromedriver-binary==141.0.7347.0.0",
  "selenium==4.34.2",
  "jinja2==3.1.6",
  "lxml==6.0.0",
  "numpy==2.2.6",
  "python-dotenv==1.1.1",
  "types-psutil==7.0.0.20250822",
  "types-PyYAML==6.0.12.20250822",
  "mcp==1.13.1",
  "fastmcp==2.11.3",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
  "pytest-aiohttp==1.0.5",
  "pytest-cov==6.2.1",
  "pytest-timeout==2.4.0",
  "pytest-mock==3.14.1",
  "faker==37.5.3",
  "ruff==0.12.8",
  "pre-commit==4.3.0",
  "mypy==1.17.1",
]

[project.scripts]
run-chrome = "browser.scripts.run_chrome:main"
run-browser-tests = "browser.scripts.run_tests:main"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["browser*"]

[tool.uv]
package = true
dev-dependencies = [
  "pytest==8.4.1",
  "pytest-asyncio==1.1.0",
  "pytest-aiohttp==1.0.5",
  "pytest-cov==6.2.1",
  "pytest-timeout==2.4.0",
  "pytest-mock==3.14.1",
  "faker==37.5.3",
  "ruff==0.12.8",
  "pre-commit==4.3.0",
  "mypy==1.17.1",
]

[tool.ruff]
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
